feat(trace): add oauth span
